[extrinsic] extracted contract xt's to submodule'
